ansible/ansible_collections/oblt/framework/roles/common/files/deployments/docker/init-policies/requirements.txt
PyYAML==6.0.1
requests==2.32.0
click==8.1.3
click-completion==0.5.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
